+++ apps/ledger-live-desktop/src/renderer/components/Onboarding/Screens/Welcome/Carousel.tsx
@@ -3,6 +3,7 @@ import React, { useState, useCallback, useEffect, useRef } from "react";
 import styled from "styled-components";
 import { TransitionGroup } from "react-transition-group";
 import { Flex, Icons, Logos, Text } from "@ledgerhq/react-ui";
+import { getEnv } from "@ledgerhq/live-common/env";
 import TransitionSlide from "@ledgerhq/react-ui/components/transitions/TransitionSlide";
 
 const Wrapper = styled(Flex).attrs({